backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
boto3==1.34.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
basicsr==1.4.2
realesrgan==0.3.0
scikit-image==0.22.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
